+++ src/main/java/de/pt400c/defaultsettings/gui/MenuScreen.java
@@ -31,7 +31,7 @@ public class MenuScreen extends Segment {
 	}
 	
 	@Override
-    public void render(float mouseX, float mouseY, float partialTicks) {
+    public void render(int mouseX, int mouseY, float partialTicks) {
 
         synchronized (this.variants) {
         	this.variants.get(this.index).render(mouseX, mouseY, partialTicks);
@@ -40,7 +40,7 @@ public class MenuScreen extends Segment {
 	}
 	
 	@Override
-	public boolean mouseClicked(double mouseX, double mouseY, int mouseButton) {
+	public boolean mouseClicked(int mouseX, int mouseY, int mouseButton) {
 		synchronized (this.variants) {
 			this.variants.get(this.index).mouseClicked(mouseX, mouseY, mouseButton);
 		}
@@ -48,7 +48,7 @@ public class MenuScreen extends Segment {
 	}
 	
 	@Override
-	public boolean mouseDragged(double p_mouseDragged_1_, double p_mouseDragged_3_, int p_mouseDragged_5_) {
+	public boolean mouseDragged(int p_mouseDragged_1_, int p_mouseDragged_3_, int p_mouseDragged_5_) {
 		synchronized (this.variants) {
 			this.variants.get(this.index).mouseDragged(p_mouseDragged_1_, p_mouseDragged_3_, p_mouseDragged_5_);
 
@@ -57,7 +57,7 @@ public class MenuScreen extends Segment {
 	}
 	
 	@Override
-	public boolean mouseScrolled(double p_mouseScrolled_1_) {
+	public boolean mouseScrolled(float p_mouseScrolled_1_) {
 		synchronized (this.variants) {
 			this.variants.get(this.index).mouseScrolled(p_mouseScrolled_1_);
 		}
@@ -79,7 +79,7 @@ public class MenuScreen extends Segment {
 	}
 	
 	@Override
-	public boolean mouseReleased(double p_mouseReleased_1_, double p_mouseReleased_3_, int p_mouseReleased_5_) {
+	public boolean mouseReleased(int p_mouseReleased_1_, int p_mouseReleased_3_, int p_mouseReleased_5_) {
 		synchronized (this.variants) {
 
 			this.variants.get(this.index).mouseReleased(p_mouseReleased_1_, p_mouseReleased_3_, p_mouseReleased_5_);
@@ -128,5 +128,4 @@ public class MenuScreen extends Segment {
     		this.supply = supply;
     	}
     }
-
 }
